fly/requirements.txt
fastapi
breez-sdk-liquid
python-dotenv
uvicorn
requests
pydantic